OGEMA 2.1.0 public release

+++ src/core/ref-impl/administration/src/main/java/org/ogema/impl/administration/BundleStoragePolicy.java
@@ -91,7 +91,7 @@ public class BundleStoragePolicy extends BundleTracker<Bundle> implements Bundle
 			log("Bundle update ", bundle, event);
 			return null;
 		}
-		int sl = (int) level.getStartLevel();
+		int sl = level.getStartLevel();
 		if (sl >= minStartLevel && !isURP) {
 			if (event.getType() == BundleEvent.INSTALLED)
 				constructStorageArea(bundle);
